+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ## [Unreleased]
 
 ### Added
+- Add check for snapshot age (to avoid useless old backup snapshots) (@VeselaHouba)
+
+### Added
 - Add sensu go bonsai asset (@danragnar)
 - Bump bundler to 2.1
 - Allow to specify the LVM command to run